Add files via upload
This is what we updated after the first one was uploaded. It runs, both pads show up in the window. But it seems that the fclose and fopen are in the wrong positions. Zack said something about those two needing to be outside of the loop to work. So maybe we can play around with that? 
The point though, is that both pads WILL show up on the school computers when you run player1 and player2, and they do! 
We also need to get the ball bouncing around and off the paddles. That has to be done by step 4.